Fix things up for cross-platform development
Replace __debugbreak with a per-platform function and define various
macros per-architecture and per-host

PLATFORM_* is now defined to 1 for the currently compiled architecture

+++ host/host_win32.cc
@@ -97,7 +97,7 @@ public:
 		std::vector<uint8_t> data;
 		data.resize((size_t)size.QuadPart);
 		DWORD n;
-		ReadFile(fd_, data.data(), data.size(), &n, nullptr);
+		ReadFile(fd_, data.data(), (DWORD)data.size(), &n, nullptr);
 
 		return data;
 	}
